pyproject.toml
[project]
name = "query-planner"
version = "0.1.0"
description = "Universal object recommendation system based on questionnaire surveys"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.0",
    "python-dotenv>=1.0",
    "jinja2>=3.0",
    "fuzzy4>=0.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=4.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/query_planner"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
